Beheer: gebruik gepubliceerde Mermaid plugin
